extract binary protocol and pg_model
