+++ src/node/src/query_engine/tests/schema.rs
@@ -13,7 +13,10 @@
 // limitations under the License.
 
 use super::*;
-use protocol::results::{QueryError, QueryEvent};
+use pg_model::{
+    results::{QueryError, QueryEvent},
+    Command,
+};
 
 #[rstest::rstest]
 fn create_schema(empty_database: (InMemory, ResultCollector)) {
