+++ src/node/src/query_engine/tests/update.rs
@@ -13,10 +13,10 @@
 // limitations under the License.
 
 use super::*;
-use pg_model::pg_types::PostgreSqlType;
-use protocol::{
-    messages::ColumnMetadata,
+use pg_model::{
+    pg_types::PostgreSqlType,
     results::{QueryError, QueryEvent},
+    Command,
 };
 
 #[rstest::rstest]
@@ -42,10 +42,9 @@ fn update_all_records(database_with_schema: (InMemory, ResultCollector)) {
         })
         .expect("query executed");
     collector.assert_receive_many(vec![
-        Ok(QueryEvent::RowDescription(vec![ColumnMetadata::new(
-            "column_test",
-            PostgreSqlType::SmallInt,
-        )])),
+        Ok(QueryEvent::RowDescription(vec![
+            PostgreSqlType::SmallInt.as_column_metadata("column_test")
+        ])),
         Ok(QueryEvent::DataRow(vec!["123".to_owned()])),
         Ok(QueryEvent::DataRow(vec!["456".to_owned()])),
         Ok(QueryEvent::RecordsSelected(2)),
@@ -64,10 +63,9 @@ fn update_all_records(database_with_schema: (InMemory, ResultCollector)) {
         })
         .expect("query executed");
     collector.assert_receive_many(vec![
-        Ok(QueryEvent::RowDescription(vec![ColumnMetadata::new(
-            "column_test",
-            PostgreSqlType::SmallInt,
-        )])),
+        Ok(QueryEvent::RowDescription(vec![
+            PostgreSqlType::SmallInt.as_column_metadata("column_test")
+        ])),
         Ok(QueryEvent::DataRow(vec!["789".to_owned()])),
         Ok(QueryEvent::DataRow(vec!["789".to_owned()])),
         Ok(QueryEvent::RecordsSelected(2)),
@@ -98,8 +96,8 @@ fn update_single_column_of_all_records(database_with_schema: (InMemory, ResultCo
         .expect("query executed");
     collector.assert_receive_many(vec![
         Ok(QueryEvent::RowDescription(vec![
-            ColumnMetadata::new("col1", PostgreSqlType::SmallInt),
-            ColumnMetadata::new("col2", PostgreSqlType::SmallInt),
+            PostgreSqlType::SmallInt.as_column_metadata("col1"),
+            PostgreSqlType::SmallInt.as_column_metadata("col2"),
         ])),
         Ok(QueryEvent::DataRow(vec!["123".to_owned(), "789".to_owned()])),
         Ok(QueryEvent::DataRow(vec!["456".to_owned(), "789".to_owned()])),
@@ -120,8 +118,8 @@ fn update_single_column_of_all_records(database_with_schema: (InMemory, ResultCo
         .expect("query executed");
     collector.assert_receive_many(vec![
         Ok(QueryEvent::RowDescription(vec![
-            ColumnMetadata::new("col1", PostgreSqlType::SmallInt),
-            ColumnMetadata::new("col2", PostgreSqlType::SmallInt),
+            PostgreSqlType::SmallInt.as_column_metadata("col1"),
+            PostgreSqlType::SmallInt.as_column_metadata("col2"),
         ])),
         Ok(QueryEvent::DataRow(vec!["123".to_owned(), "357".to_owned()])),
         Ok(QueryEvent::DataRow(vec!["456".to_owned(), "357".to_owned()])),
@@ -153,9 +151,9 @@ fn update_multiple_columns_of_all_records(database_with_schema: (InMemory, Resul
         .expect("query executed");
     collector.assert_receive_many(vec![
         Ok(QueryEvent::RowDescription(vec![
-            ColumnMetadata::new("col1", PostgreSqlType::SmallInt),
-            ColumnMetadata::new("col2", PostgreSqlType::SmallInt),
-            ColumnMetadata::new("col3", PostgreSqlType::SmallInt),
+            PostgreSqlType::SmallInt.as_column_metadata("col1"),
+            PostgreSqlType::SmallInt.as_column_metadata("col2"),
+            PostgreSqlType::SmallInt.as_column_metadata("col3"),
         ])),
         Ok(QueryEvent::DataRow(vec![
             "111".to_owned(),
@@ -185,9 +183,9 @@ fn update_multiple_columns_of_all_records(database_with_schema: (InMemory, Resul
 
     collector.assert_receive_many(vec![
         Ok(QueryEvent::RowDescription(vec![
-            ColumnMetadata::new("col1", PostgreSqlType::SmallInt),
-            ColumnMetadata::new("col2", PostgreSqlType::SmallInt),
-            ColumnMetadata::new("col3", PostgreSqlType::SmallInt),
+            PostgreSqlType::SmallInt.as_column_metadata("col1"),
+            PostgreSqlType::SmallInt.as_column_metadata("col2"),
+            PostgreSqlType::SmallInt.as_column_metadata("col3"),
         ])),
         Ok(QueryEvent::DataRow(vec![
             "999".to_owned(),
@@ -228,9 +226,9 @@ fn update_all_records_in_multiple_columns(database_with_schema: (InMemory, Resul
         .expect("query executed");
     collector.assert_receive_many(vec![
         Ok(QueryEvent::RowDescription(vec![
-            ColumnMetadata::new("column_1", PostgreSqlType::SmallInt),
-            ColumnMetadata::new("column_2", PostgreSqlType::SmallInt),
-            ColumnMetadata::new("column_3", PostgreSqlType::SmallInt),
+            PostgreSqlType::SmallInt.as_column_metadata("column_1"),
+            PostgreSqlType::SmallInt.as_column_metadata("column_2"),
+            PostgreSqlType::SmallInt.as_column_metadata("column_3"),
         ])),
         Ok(QueryEvent::DataRow(vec![
             "1".to_owned(),
@@ -265,9 +263,9 @@ fn update_all_records_in_multiple_columns(database_with_schema: (InMemory, Resul
 
     collector.assert_receive_many(vec![
         Ok(QueryEvent::RowDescription(vec![
-            ColumnMetadata::new("column_1", PostgreSqlType::SmallInt),
-            ColumnMetadata::new("column_2", PostgreSqlType::SmallInt),
-            ColumnMetadata::new("column_3", PostgreSqlType::SmallInt),
+            PostgreSqlType::SmallInt.as_column_metadata("column_1"),
+            PostgreSqlType::SmallInt.as_column_metadata("column_2"),
+            PostgreSqlType::SmallInt.as_column_metadata("column_3"),
         ])),
         Ok(QueryEvent::DataRow(vec![
             "10".to_owned(),
@@ -322,10 +320,9 @@ fn update_non_existent_columns_of_records(database_with_schema: (InMemory, Resul
         })
         .expect("query executed");
     collector.assert_receive_many(vec![
-        Ok(QueryEvent::RowDescription(vec![ColumnMetadata::new(
-            "column_test",
-            PostgreSqlType::SmallInt,
-        )])),
+        Ok(QueryEvent::RowDescription(vec![
+            PostgreSqlType::SmallInt.as_column_metadata("column_test")
+        ])),
         Ok(QueryEvent::DataRow(vec!["123".to_owned()])),
         Ok(QueryEvent::RecordsSelected(1)),
     ]);
@@ -370,9 +367,9 @@ fn test_update_with_dynamic_expression(database_with_schema: (InMemory, ResultCo
         .expect("query executed");
     collector.assert_receive_many(vec![
         Ok(QueryEvent::RowDescription(vec![
-            ColumnMetadata::new("si_column_1", PostgreSqlType::SmallInt),
-            ColumnMetadata::new("si_column_2", PostgreSqlType::SmallInt),
-            ColumnMetadata::new("si_column_3", PostgreSqlType::SmallInt),
+            PostgreSqlType::SmallInt.as_column_metadata("si_column_1"),
+            PostgreSqlType::SmallInt.as_column_metadata("si_column_2"),
+            PostgreSqlType::SmallInt.as_column_metadata("si_column_3"),
         ])),
         Ok(QueryEvent::DataRow(vec![
             "1".to_owned(),
@@ -411,9 +408,9 @@ fn test_update_with_dynamic_expression(database_with_schema: (InMemory, ResultCo
         .expect("query executed");
     collector.assert_receive_many(vec![
         Ok(QueryEvent::RowDescription(vec![
-            ColumnMetadata::new("si_column_1", PostgreSqlType::SmallInt),
-            ColumnMetadata::new("si_column_2", PostgreSqlType::SmallInt),
-            ColumnMetadata::new("si_column_3", PostgreSqlType::SmallInt),
+            PostgreSqlType::SmallInt.as_column_metadata("si_column_1"),
+            PostgreSqlType::SmallInt.as_column_metadata("si_column_2"),
+            PostgreSqlType::SmallInt.as_column_metadata("si_column_3"),
         ])),
         Ok(QueryEvent::DataRow(vec![
             (2 * 1).to_string(),
@@ -485,10 +482,9 @@ mod operators {
                     })
                     .expect("query executed");
                 collector.assert_receive_many(vec![
-                    Ok(QueryEvent::RowDescription(vec![ColumnMetadata::new(
-                        "column_si",
-                        PostgreSqlType::SmallInt,
-                    )])),
+                    Ok(QueryEvent::RowDescription(vec![
+                        PostgreSqlType::SmallInt.as_column_metadata("column_si")
+                    ])),
                     Ok(QueryEvent::DataRow(vec!["3".to_owned()])),
                     Ok(QueryEvent::RecordsSelected(1)),
                 ]);
@@ -511,10 +507,9 @@ mod operators {
                     .expect("query executed");
 
                 collector.assert_receive_many(vec![
-                    Ok(QueryEvent::RowDescription(vec![ColumnMetadata::new(
-                        "column_si",
-                        PostgreSqlType::SmallInt,
-                    )])),
+                    Ok(QueryEvent::RowDescription(vec![
+                        PostgreSqlType::SmallInt.as_column_metadata("column_si")
+                    ])),
                     Ok(QueryEvent::DataRow(vec!["-1".to_owned()])),
                     Ok(QueryEvent::RecordsSelected(1)),
                 ]);
@@ -537,10 +532,9 @@ mod operators {
                     .expect("query executed");
 
                 collector.assert_receive_many(vec![
-                    Ok(QueryEvent::RowDescription(vec![ColumnMetadata::new(
-                        "column_si",
-                        PostgreSqlType::SmallInt,
-                    )])),
+                    Ok(QueryEvent::RowDescription(vec![
+                        PostgreSqlType::SmallInt.as_column_metadata("column_si")
+                    ])),
                     Ok(QueryEvent::DataRow(vec!["6".to_owned()])),
                     Ok(QueryEvent::RecordsSelected(1)),
                 ]);
@@ -562,10 +556,9 @@ mod operators {
                     })
                     .expect("query executed");
                 collector.assert_receive_many(vec![
-                    Ok(QueryEvent::RowDescription(vec![ColumnMetadata::new(
-                        "column_si",
-                        PostgreSqlType::SmallInt,
-                    )])),
+                    Ok(QueryEvent::RowDescription(vec![
+                        PostgreSqlType::SmallInt.as_column_metadata("column_si")
+                    ])),
                     Ok(QueryEvent::DataRow(vec!["4".to_owned()])),
                     Ok(QueryEvent::RecordsSelected(1)),
                 ]);
@@ -587,10 +580,9 @@ mod operators {
                     })
                     .expect("query executed");
                 collector.assert_receive_many(vec![
-                    Ok(QueryEvent::RowDescription(vec![ColumnMetadata::new(
-                        "column_si",
-                        PostgreSqlType::SmallInt,
-                    )])),
+                    Ok(QueryEvent::RowDescription(vec![
+                        PostgreSqlType::SmallInt.as_column_metadata("column_si")
+                    ])),
                     Ok(QueryEvent::DataRow(vec!["0".to_owned()])),
                     Ok(QueryEvent::RecordsSelected(1)),
                 ]);
@@ -615,10 +607,9 @@ mod operators {
                     })
                     .expect("query executed");
                 collector.assert_receive_many(vec![
-                    Ok(QueryEvent::RowDescription(vec![ColumnMetadata::new(
-                        "column_si",
-                        PostgreSqlType::SmallInt,
-                    )])),
+                    Ok(QueryEvent::RowDescription(vec![
+                        PostgreSqlType::SmallInt.as_column_metadata("column_si")
+                    ])),
                     Ok(QueryEvent::DataRow(vec!["64".to_owned()])),
                     Ok(QueryEvent::RecordsSelected(1)),
                 ]);
@@ -642,10 +633,9 @@ mod operators {
                     })
                     .expect("query executed");
                 collector.assert_receive_many(vec![
-                    Ok(QueryEvent::RowDescription(vec![ColumnMetadata::new(
-                        "column_si",
-                        PostgreSqlType::SmallInt,
-                    )])),
+                    Ok(QueryEvent::RowDescription(vec![
+                        PostgreSqlType::SmallInt.as_column_metadata("column_si")
+                    ])),
                     Ok(QueryEvent::DataRow(vec!["4".to_owned()])),
                     Ok(QueryEvent::RecordsSelected(1)),
                 ]);
@@ -669,10 +659,9 @@ mod operators {
                     })
                     .expect("query executed");
                 collector.assert_receive_many(vec![
-                    Ok(QueryEvent::RowDescription(vec![ColumnMetadata::new(
-                        "column_si",
-                        PostgreSqlType::SmallInt,
-                    )])),
+                    Ok(QueryEvent::RowDescription(vec![
+                        PostgreSqlType::SmallInt.as_column_metadata("column_si")
+                    ])),
                     Ok(QueryEvent::DataRow(vec!["2".to_owned()])),
                     Ok(QueryEvent::RecordsSelected(1)),
                 ]);
@@ -696,10 +685,9 @@ mod operators {
                     })
                     .expect("query executed");
                 collector.assert_receive_many(vec![
-                    Ok(QueryEvent::RowDescription(vec![ColumnMetadata::new(
-                        "column_si",
-                        PostgreSqlType::SmallInt,
-                    )])),
+                    Ok(QueryEvent::RowDescription(vec![
+                        PostgreSqlType::SmallInt.as_column_metadata("column_si")
+                    ])),
                     Ok(QueryEvent::DataRow(vec!["120".to_owned()])),
                     Ok(QueryEvent::RecordsSelected(1)),
                 ]);
@@ -723,10 +711,9 @@ mod operators {
                     })
                     .expect("query executed");
                 collector.assert_receive_many(vec![
-                    Ok(QueryEvent::RowDescription(vec![ColumnMetadata::new(
-                        "column_si",
-                        PostgreSqlType::SmallInt,
-                    )])),
+                    Ok(QueryEvent::RowDescription(vec![
+                        PostgreSqlType::SmallInt.as_column_metadata("column_si")
+                    ])),
                     Ok(QueryEvent::DataRow(vec!["120".to_owned()])),
                     Ok(QueryEvent::RecordsSelected(1)),
                 ]);
@@ -750,10 +737,9 @@ mod operators {
                     })
                     .expect("query executed");
                 collector.assert_receive_many(vec![
-                    Ok(QueryEvent::RowDescription(vec![ColumnMetadata::new(
-                        "column_si",
-                        PostgreSqlType::SmallInt,
-                    )])),
+                    Ok(QueryEvent::RowDescription(vec![
+                        PostgreSqlType::SmallInt.as_column_metadata("column_si")
+                    ])),
                     Ok(QueryEvent::DataRow(vec!["5".to_owned()])),
                     Ok(QueryEvent::RecordsSelected(1)),
                 ]);
@@ -775,10 +761,9 @@ mod operators {
                     })
                     .expect("query executed");
                 collector.assert_receive_many(vec![
-                    Ok(QueryEvent::RowDescription(vec![ColumnMetadata::new(
-                        "column_si",
-                        PostgreSqlType::SmallInt,
-                    )])),
+                    Ok(QueryEvent::RowDescription(vec![
+                        PostgreSqlType::SmallInt.as_column_metadata("column_si")
+                    ])),
                     Ok(QueryEvent::DataRow(vec!["1".to_owned()])),
                     Ok(QueryEvent::RecordsSelected(1)),
                 ]);
@@ -800,10 +785,9 @@ mod operators {
                     })
                     .expect("query executed");
                 collector.assert_receive_many(vec![
-                    Ok(QueryEvent::RowDescription(vec![ColumnMetadata::new(
-                        "column_si",
-                        PostgreSqlType::SmallInt,
-                    )])),
+                    Ok(QueryEvent::RowDescription(vec![
+                        PostgreSqlType::SmallInt.as_column_metadata("column_si")
+                    ])),
                     Ok(QueryEvent::DataRow(vec!["7".to_owned()])),
                     Ok(QueryEvent::RecordsSelected(1)),
                 ]);
@@ -827,10 +811,9 @@ mod operators {
                     })
                     .expect("query executed");
                 collector.assert_receive_many(vec![
-                    Ok(QueryEvent::RowDescription(vec![ColumnMetadata::new(
-                        "column_si",
-                        PostgreSqlType::SmallInt,
-                    )])),
+                    Ok(QueryEvent::RowDescription(vec![
+                        PostgreSqlType::SmallInt.as_column_metadata("column_si")
+                    ])),
                     Ok(QueryEvent::DataRow(vec!["-2".to_owned()])),
                     Ok(QueryEvent::RecordsSelected(1)),
                 ]);
@@ -854,10 +837,9 @@ mod operators {
                     })
                     .expect("query executed");
                 collector.assert_receive_many(vec![
-                    Ok(QueryEvent::RowDescription(vec![ColumnMetadata::new(
-                        "column_si",
-                        PostgreSqlType::SmallInt,
-                    )])),
+                    Ok(QueryEvent::RowDescription(vec![
+                        PostgreSqlType::SmallInt.as_column_metadata("column_si")
+                    ])),
                     Ok(QueryEvent::DataRow(vec!["16".to_owned()])),
                     Ok(QueryEvent::RecordsSelected(1)),
                 ]);
@@ -881,10 +863,9 @@ mod operators {
                     })
                     .expect("query executed");
                 collector.assert_receive_many(vec![
-                    Ok(QueryEvent::RowDescription(vec![ColumnMetadata::new(
-                        "column_si",
-                        PostgreSqlType::SmallInt,
-                    )])),
+                    Ok(QueryEvent::RowDescription(vec![
+                        PostgreSqlType::SmallInt.as_column_metadata("column_si")
+                    ])),
                     Ok(QueryEvent::DataRow(vec!["2".to_owned()])),
                     Ok(QueryEvent::RecordsSelected(1)),
                 ]);
@@ -906,10 +887,9 @@ mod operators {
                     })
                     .expect("query executed");
                 collector.assert_receive_many(vec![
-                    Ok(QueryEvent::RowDescription(vec![ColumnMetadata::new(
-                        "column_si",
-                        PostgreSqlType::SmallInt,
-                    )])),
+                    Ok(QueryEvent::RowDescription(vec![
+                        PostgreSqlType::SmallInt.as_column_metadata("column_si")
+                    ])),
                     Ok(QueryEvent::DataRow(vec!["5".to_owned()])),
                     Ok(QueryEvent::RecordsSelected(1)),
                 ]);
@@ -960,10 +940,9 @@ mod operators {
                 })
                 .expect("query executed");
             collector.assert_receive_many(vec![
-                Ok(QueryEvent::RowDescription(vec![ColumnMetadata::new(
-                    "strings",
-                    PostgreSqlType::Char,
-                )])),
+                Ok(QueryEvent::RowDescription(vec![
+                    PostgreSqlType::Char.as_column_metadata("strings")
+                ])),
                 Ok(QueryEvent::DataRow(vec!["12345".to_owned()])),
                 Ok(QueryEvent::RecordsSelected(1)),
             ]);
@@ -985,10 +964,9 @@ mod operators {
                 })
                 .expect("query executed");
             collector.assert_receive_many(vec![
-                Ok(QueryEvent::RowDescription(vec![ColumnMetadata::new(
-                    "strings",
-                    PostgreSqlType::Char,
-                )])),
+                Ok(QueryEvent::RowDescription(vec![
+                    PostgreSqlType::Char.as_column_metadata("strings")
+                ])),
                 Ok(QueryEvent::DataRow(vec!["145".to_owned()])),
                 Ok(QueryEvent::RecordsSelected(1)),
             ]);
@@ -1006,10 +984,9 @@ mod operators {
                 })
                 .expect("query executed");
             collector.assert_receive_many(vec![
-                Ok(QueryEvent::RowDescription(vec![ColumnMetadata::new(
-                    "strings",
-                    PostgreSqlType::Char,
-                )])),
+                Ok(QueryEvent::RowDescription(vec![
+                    PostgreSqlType::Char.as_column_metadata("strings")
+                ])),
                 Ok(QueryEvent::DataRow(vec!["451".to_owned()])),
                 Ok(QueryEvent::RecordsSelected(1)),
             ]);
