+++ src/node/src/query_engine/tests/where_clause.rs
@@ -13,8 +13,7 @@
 // limitations under the License.
 
 use super::*;
-use pg_model::pg_types::PostgreSqlType;
-use protocol::{messages::ColumnMetadata, results::QueryEvent};
+use pg_model::{pg_types::PostgreSqlType, results::QueryEvent, Command};
 
 #[rstest::rstest]
 fn select_row_by_column_equality_predicate(database_with_schema: (InMemory, ResultCollector)) {
@@ -41,9 +40,9 @@ fn select_row_by_column_equality_predicate(database_with_schema: (InMemory, Resu
         .expect("query executed");
     collector.assert_receive_many(vec![
         Ok(QueryEvent::RowDescription(vec![
-            ColumnMetadata::new("column_1", PostgreSqlType::SmallInt),
-            ColumnMetadata::new("column_2", PostgreSqlType::SmallInt),
-            ColumnMetadata::new("column_3", PostgreSqlType::SmallInt),
+            PostgreSqlType::SmallInt.as_column_metadata("column_1"),
+            PostgreSqlType::SmallInt.as_column_metadata("column_2"),
+            PostgreSqlType::SmallInt.as_column_metadata("column_3"),
         ])),
         Ok(QueryEvent::DataRow(vec![
             "1".to_owned(),
