+++ src/pg_model/Cargo.toml
@@ -6,9 +6,21 @@ edition = "2018"
 publish = false
 
 [dependencies]
-bigdecimal = "0.2.0"
+protocol = { path = "../protocol" }
+
+async-dup = "1.2.1"
+async-mutex = "1.4.0"
+#TODO temporal to have ability to run ./local/code_coverage.sh
+async-native-tls = { git = "https://github.com/alex-dukhno/async-native-tls.git", branch = "new-native-tls" }
+async-trait = "0.1.40"
+blocking = "1.0.0"
+bigdecimal = { version = "0.2.0", features = ["string-only"] }
 byteorder = "1.3.4"
 sqlparser = { git = "https://github.com/ballista-compute/sqlparser-rs.git", branch = "main", features = ["bigdecimal"] }
+futures-lite = "1.10.0"
+log = "0.4.11"
+rand = "0.7"
+tempfile = "3.1.0"
 
 [dev-dependencies]
 rstest = "0.6.4"
