+++ src/query_planner/Cargo.toml
@@ -13,7 +13,7 @@ metadata = { path = "../metadata" }
 plan = { path = "../plan" }
 sql_model = { path = "../sql_model" }
 
-bigdecimal = "0.2.0"
+bigdecimal = { version = "0.2.0", features = ["string-only"] }
 sqlparser = { git = "https://github.com/ballista-compute/sqlparser-rs.git", branch = "main", features = ["bigdecimal"] }
 
 [dev-dependencies]
