+++ src/query_planner/src/tests/select.rs
@@ -20,7 +20,7 @@ use sqlparser::ast::{ObjectName, Query, Select, SelectItem, SetExpr, Statement, 
 fn select_from_table_that_in_nonexistent_schema(planner: QueryPlanner) {
     assert_eq!(
         planner.plan(&Statement::Query(Box::new(Query {
-            ctes: vec![],
+            with: None,
             body: SetExpr::Select(Box::new(Select {
                 distinct: false,
                 top: None,
@@ -51,7 +51,7 @@ fn select_from_table_that_in_nonexistent_schema(planner: QueryPlanner) {
 fn select_from_nonexistent_table(planner_with_schema: QueryPlanner) {
     assert_eq!(
         planner_with_schema.plan(&Statement::Query(Box::new(Query {
-            ctes: vec![],
+            with: None,
             body: SetExpr::Select(Box::new(Select {
                 distinct: false,
                 top: None,
@@ -85,7 +85,7 @@ fn select_from_nonexistent_table(planner_with_schema: QueryPlanner) {
 fn select_from_table_with_unqualified_name(planner_with_schema: QueryPlanner) {
     assert_eq!(
         planner_with_schema.plan(&Statement::Query(Box::new(Query {
-            ctes: vec![],
+            with: None,
             body: SetExpr::Select(Box::new(Select {
                 distinct: false,
                 top: None,
@@ -118,7 +118,7 @@ fn select_from_table_with_unqualified_name(planner_with_schema: QueryPlanner) {
 fn select_from_table_with_unsupported_name(planner_with_schema: QueryPlanner) {
     assert_eq!(
         planner_with_schema.plan(&Statement::Query(Box::new(Query {
-            ctes: vec![],
+            with: None,
             body: SetExpr::Select(Box::new(Select {
                 distinct: false,
                 top: None,
@@ -156,7 +156,7 @@ fn select_from_table_with_unsupported_name(planner_with_schema: QueryPlanner) {
 fn select_from_table(planner_with_no_column_table: QueryPlanner) {
     assert_eq!(
         planner_with_no_column_table.plan(&Statement::Query(Box::new(Query {
-            ctes: vec![],
+            with: None,
             body: SetExpr::Select(Box::new(Select {
                 distinct: false,
                 top: None,
